requirements.txt
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langgraph==0.0.28
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pypdf==3.17.1
chromadb==0.4.22
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic==2.5.3
tiktoken==0.5.2
tenacity==8.2.3
prometheus-client==0.19.0
SQLAlchemy==2.0.23
langfuse